final version

+++ src/main/scala/util.scala
@@ -9,15 +9,16 @@ object util {
   }
 
   def toVector(row: Row) = Vectors.dense(Array(
-    row.getAs[Int](2).toDouble, // Pclass
-    row.getAs[Double](3),       // Age
-    row.getAs[Int](4).toDouble, // SibSp
-    row.getAs[Int](5).toDouble, // Parch
-    row.getAs[Double](6),       // Fare
-    row.getAs[Double](7),       // Name_discrete
-    row.getAs[Double](8),       // Sex_discrete
-    row.getAs[Double](9),       // Ticket_discrete
-    row.getAs[Double](10),      // Cabin_discrete
-    row.getAs[Double](11)))     // Embarked_discrete
+    row.getAs[Int]   (2).toDouble, // Pclass
+    row.getAs[Double](3),          // Age
+    row.getAs[Double](4),          // Fare
+    row.getAs[Int]   (5).toDouble, // FamilySize
+    row.getAs[Int]   (6),          // NumCabins
+    row.getAs[Double](7),         // Sex_discrete
+    row.getAs[Double](8),         // Embarked_discrete
+    row.getAs[Double](9),         // Title_discrete
+    row.getAs[Double](10),         // TicketCode_discrete
+    row.getAs[Double](11)          // Deck_discrete
+  ))
 
 }
